Untitled...
A subjective visual representation of the working mind of a programmer.
Connecting the loose association found in the surreal to semantic error
occurring from miscommunication between man and machine as oppose to man
and its surrounding world. The miscommunication displayed does not only
represent the communication difficulties of man and machine but any
difficulty of communication between minds that operate differently or
any form of miscommunication there is.